Refactor OCEs for Summer19 study into sep. files
