+++ imports/api/Testing/testingconstants.js
@@ -3,10 +3,9 @@ import { Meteor } from "meteor/meteor";
 import { Submissions } from "../OCEManager/currentNeeds";
 
 import { addContribution } from '../OCEManager/OCEs/methods';
-import {getDetectorUniqueKey} from "../UserMonitor/detectors/methods";
 import {notify, notifyUsersInIncident, notifyUsersInNeed} from "../OpportunisticCoordinator/server/noticationMethods";
 import {Incidents} from "../OCEManager/OCEs/experiences";
-import {addStaticAffordanceToNeeds} from "./oce_api_helpers";
+import {getDetectorUniqueKey, addStaticAffordanceToNeeds} from "./oce_api_helpers";
 
 import {LOCATIONS} from "./LOCATIONS";
 import {USERS} from "./USERS";
@@ -195,216 +194,7 @@ function createStorytime(version) {
   };
 }
 
-/**
- * createStorytimeImproved (improved on May 30th 2019)
- *
- * @param knowsGroup (e.g., "knowsOlin")
- * @param version (e.g., 0 determines a starting point in the story of 0)
- * @return {{contributionTypes: {notificationDelay: number, numberNeeded: number, needName: string, toPass: {firstSentence: string, instruction: string, dropdownChoices: {name: string, options: *}, situation}, situation: {number: string, detector}}[], participateTemplate: string, notificationText: string, name: string, resultsTemplate: string, description: string, callbacks: *[], _id: *}}
- */
-const createStorytimeImproved = (group, version) => {
-  const knowsGroup = `knows${group}`;
 
-  // setup places and detectors for storytime
-  let places = ["niceish_day", "bar", "train", "forest", "restaurant", "gym"];
-  let detectorUniqueKeys = places.map((x) => { return Random.id(); });
-  let detectorNames = [];
-  let dropdownText = [
-    'Swirling Clouds',
-    'Drinking butterbeer',
-    'Hogwarts Express',
-    'Forbidden Forest',
-    'Dinner at the Great Hall',
-    'Training in the Room of Requirement',
-  ];
-  const notificationSubjects = [
-    'Is the weather cloudy today?',
-    'Drinking at a bar?',
-    'Waiting at a train station?',
-    'Are you at a park?',
-    'Eating out at a restaurant?',
-    'Working out at the gym?'
-  ];
-  const notificationTexts = [
-    'Create a story with others using the cloudy scene above you',
-    'Create a story with others using the bar scene around you',
-    'Create a story with others using the train station scene around you',
-    'Create a story with others using the park scene around you',
-    'Create a story with others using the dining scene around you',
-    'Create a story with others using the gym scene around you',
-  ];
-
-  // updating detector rules to have static affordances
-  _.forEach(places, (place, i) => {
-    let newVars = JSON.parse(JSON.stringify(DETECTORS[place]['variables']));
-    newVars.push(`var ${knowsGroup};`);
-    let newRules = JSON.parse(JSON.stringify(DETECTORS[place]['rules']));
-    // modify last detector rule
-    // when rules has a flat structure where rules.length == 1, last rule is the predicate
-    // i.e. ['(diners || restaurants || cafeteria || food_court);']
-    // when rules have a nested structure where rules.length > 1, last rule is the predicate
-    // i.e. ['worship_places = (buddhist_temples || churches);', '(worship_places || landmarks);']
-    let lastRule = newRules.pop();
-    // each rule has a `;` at end, i.e. (rain && park);
-    // in order to modify the rule, must add predicate preceding the rule
-    let lastRuleNoSemicolon = lastRule.split(';')[0];
-    lastRule = `(${knowsGroup} && (${lastRuleNoSemicolon}))`;
-    newRules.push(lastRule);
-
-    let detectorName = `${place}_storytime_${knowsGroup}`;
-    detectorNames.push(detectorName);
-
-    DETECTORS[detectorName] = {
-      '_id': detectorUniqueKeys[i],
-      'description': `${DETECTORS[place].description} storytime ${knowsGroup}`,
-      'variables': newVars,
-      'rules': newRules
-    };
-  });
-
-  // Don't assume the Random detectorUniqueKeys we created actually exist
-  detectorUniqueKeys = detectorNames.map((name) => { return getDetectorUniqueKey(DETECTORS[name]); });
-  let DROPDOWN_OPTIONS = _.zip(dropdownText, detectorUniqueKeys);
-  let NOTIF_SUBJECT_OPTIONS = _.zip(notificationSubjects, detectorUniqueKeys);
-  let NOTIF_TEXT_OPTIONS = _.zip(notificationTexts, detectorUniqueKeys);
-
-  // create story starting point
-  let sentences = [
-    'Ron looked up at the clouds swirling above him.',
-    'Hermoine looked into her goblet, hardly realizing the unusual color of the concoction she was being forced to drink.',
-    'Harry prepared himself for a lunge, and then dove forward towards the Platform 9 3/4 wall.',
-    'The wizard looked down at their feet, hardly believing the magical plants growing in the Forbidden Forest.',
-    'Any young wizard who has their first meal in the Hogwarts Great Hall has to be surprised by the type of food on the menu.',
-    'The new wizard of Dumbledore\'s Army was training very hard in the Room of Requirement.'
-  ];
-  let firstSentence = sentences[version];
-  let [firstSituation, firstDetector] = DROPDOWN_OPTIONS[version];
-  let [firstNotificationSubject, _1] = NOTIF_SUBJECT_OPTIONS[version];
-  let [firstNotificationText, _2] = NOTIF_TEXT_OPTIONS[version];
-  // notify users when story is complete
-  let sendNotification = function (sub) {
-    let uids = Submissions.find({iid: sub.iid}).fetch().map(function (x) {
-      return x.uid;
-    });
-
-    notify(uids, sub.iid, 'Our story is finally complete. Click here to read it!',
-      '', '/apicustomresults/' + sub.iid + '/' + sub.eid);
-  };
-
-  /**
-   * NOTE: if callback depends on any variables defined outside of its scope, we must use some solution so that
-   * the variables values are substituted into the callback.toString()
-   *
-   * For a dynamic code generation solution,
-   * @see https://stackoverflow.com/questions/29182244/convert-a-string-to-a-template-string
-   * @see https://medium.com/@oprearocks/serializing-object-methods-using-es6-template-strings-and-eval-c77c894651f0
-   * @param sub
-   */
-  const storytimeCallbackImproved = function (sub) {
-    // coordinator/strategizer handles repitition in experiences
-
-    // set affordances for storytime
-    let affordance = sub.content.affordance;
-
-    // HACKY TEMPLATE DYNAMIC CODE GENERATION
-    let options = eval('${JSON.stringify(DROPDOWN_OPTIONS)}');
-    let notification_subject_options = eval('${JSON.stringify(NOTIF_SUBJECT_OPTIONS)}');
-    let notification_text_options = eval('${JSON.stringify(NOTIF_TEXT_OPTIONS)}');
-
-    let [situation, detectorUniqueKey] = options.find(function(x) {
-      return x[1] === affordance;
-    });
-    let [notificationSubject, _1] = notification_subject_options.find(function(x) {
-      return x[1] === affordance;
-    });
-    let [notificationText, _2] = notification_text_options.find(function(x) {
-      return x[1] === affordance;
-    });
-
-
-    // options = options.filter(function (x) {
-    //   return x[1] !== affordance;
-    // });
-
-    // add need if not all pages are done
-    let needName = 'page' + Random.id(3);
-    if (cb.numberOfSubmissions() === 15) {
-      needName = 'pageFinal'
-    }
-
-    // create and add contribution
-    let contribution = {
-      needName: needName,
-      notificationSubject: notificationSubject,
-      notificationText: notificationText,
-      situation: {
-        detector: affordance,
-        number: '1'
-      },
-      toPass: {
-        instruction: sub.content.sentence,
-        situation: situation,
-        previousUserId: sub.uid,
-        dropdownChoices: {
-          name: 'affordance',
-          options: options
-        }
-      },
-      numberNeeded: 1,
-      notificationDelay: 90
-    };
-
-    addContribution(sub.iid, contribution);
-  };
-
-  // FIXME(rlouie): Can't have more than version 0,1,2
-  let exp_names = [
-    "A Ron Weasley Story",
-    "A Hermoine Granger Story",
-    "A Harry Potter Story"
-  ];
-
-  // create and return storytime experience
-  return {
-    _id: Random.id(),
-    name: exp_names[version],
-    group: group,
-    participateTemplate: 'storyPage',
-    resultsTemplate: 'storybook',
-    repeatContributionsToExperienceAfterN: 1, // let one people participate before I can again
-    contributionTypes: [{
-      needName: 'pageOne',
-      notificationSubject: firstNotificationSubject,
-      notificationText: firstNotificationText,
-      situation: {
-        detector: firstDetector,
-        number: '1'
-      },
-      toPass: {
-        instruction: firstSentence,
-        firstSentence: firstSentence,
-        situation: firstSituation,
-        dropdownChoices: {
-          name: 'affordance',
-          options: DROPDOWN_OPTIONS
-        }
-      },
-      numberNeeded: 1,
-      notificationDelay: 90
-    }],
-    description: "Create a story with others, told through people's everyday surroundings",
-    callbacks: [
-      {
-        trigger: `cb.newSubmission() && (cb.numberOfSubmissions() <= 15)`,
-        // substitute any variables used outside of the callback function scope
-        function: eval('`' + storytimeCallbackImproved.toString() + '`'),
-      },
-      {
-        trigger: 'cb.incidentFinished()',
-        function: sendNotification.toString()
-      }]
-  };
-};
 
 const createIndependentStorybook = () => {
 
@@ -2363,665 +2153,8 @@ let EXPERIENCES = {
   // }
 };
 
-let CHI20_DTR_EXPERIENCES =  {
-  storytime_knowsDTR: createStorytimeImproved('DTR', 1), // version 1 starts with bars
-  halfhalf_sunny_knowsDTR: {
-    _id: Random.id(),
-    name: 'Hand Silhouette',
-    group: 'DTR',
-    participateTemplate: 'halfhalfParticipate',
-    resultsTemplate: 'halfhalfResults',
-    contributionTypes: addStaticAffordanceToNeeds('knowsDTR', [{
-      // needName MUST have structure "My Need Name XYZ"
-      needName: 'Hand Silhouette 1',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.sunny),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'Is the <span style="color: #0351ff">weather clear and sunny</span> where you are? Take a photo, <span style="color: #0351ff">holding your hand towards the sky, covering the sun.</span>',
-        exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-hands-in-front.jpg'
-      },
-      numberNeeded: 2,
-      numberAllowedToParticipateAtSameTime: 1,
-      notificationDelay: 1,
-    }]),
-    description: 'Use the sun to make a silhouette of your hand',
-    notificationText: 'View this and other available experiences',
-    callbacks: [{
-      trigger: '(cb.numberOfSubmissions() % 2) === 0',
-      function: halfhalfRespawnAndNotify('A hand silhouette was completed','View the photo').toString()
-    }]
-  },
-  halfhalf_grocery_knowsDTR: {
-    _id: Random.id(),
-    name: 'Grocery Buddies',
-    group: 'DTR',
-    participateTemplate: 'halfhalfParticipate',
-    resultsTemplate: 'halfhalfResults',
-    contributionTypes: addStaticAffordanceToNeeds('knowsDTR', [{
-      // needName MUST have structure "My Need Name XYZ"
-      needName: 'Grocery Buddies 1',
-      notificationSubject: 'Inside a grocery store?',
-      notificationText: 'Share an experience with others who are also grocery shopping',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.grocery),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'Are you at the <span style="color: #0351ff">grocery store</span>? Take a photo, <span style="color: #0351ff">holding a fruit or vegetable</span> outstretched with your hands.',
-        exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-fruit-in-hand.jpg'
-      },
-      numberNeeded: 2,
-      numberAllowedToParticipateAtSameTime: 1,
-      notificationDelay: 90,
-    }]),
-    description: 'While shopping for groceries, create a half half photo.',
-    notificationText: 'View this and other available experiences',
-    callbacks: [{
-      trigger: '(cb.numberOfSubmissions() % 2) === 0',
-      function: halfhalfRespawnAndNotify('A Grocery Buddies photo completed','View the photo').toString()
-    }]
-  },
-  halfhalf_bar_knowsDTR: {
-    _id: Random.id(),
-    name: 'Cheers',
-    group: 'DTR',
-    participateTemplate: 'halfhalfParticipate',
-    resultsTemplate: 'halfhalfResults',
-    contributionTypes: addStaticAffordanceToNeeds('knowsDTR', [{
-      // needName MUST have structure "My Need Name XYZ"
-      needName: 'Cheers 1',
-      notificationSubject: 'Drinking at a bar?',
-      notificationText: 'Share an experience with others who are also drinking at a bar',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.bar),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'What are you <span style="color: #0351ff">drinking at the bar</span>? Take a photo, while <span style="color: #0351ff">raising your glass or bottle</span> in front of you.',
-        exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-cheers.jpg'
-      },
-      numberNeeded: 2,
-      numberAllowedToParticipateAtSameTime: 1,
-      notificationDelay: 90
-    }]),
-    description: 'While enjoying your drink, create a half half photo.',
-    notificationText: 'View this and other available experiences',
-    callbacks: [{
-      trigger: '(cb.numberOfSubmissions() % 2) === 0',
-      function: halfhalfRespawnAndNotify('A Cheers photo completed','View the photo').toString()
-    }]
-  },
-  halfhalf_religious_knowsDTR: {
-    _id: Random.id(),
-    name: 'Religious Architecture',
-    group: 'DTR',
-    participateTemplate: 'halfhalfParticipate',
-    resultsTemplate: 'halfhalfResults',
-    contributionTypes: addStaticAffordanceToNeeds('knowsDTR', [{
-      // needName MUST have structure "My Need Name XYZ"
-      needName: 'Religious Architecture 1',
-      notificationSubject: 'Visiting a place of worship?',
-      notificationText: 'Share an experience with others who are also visiting a place of worship',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.castle),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'Do you notice the <span style="color: #0351ff">details of the religious building</span> near you? Do so now, by outstretching your hand and pointing out of the elements that stick out to you most.',
-        exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-religious-building.jpg'
-      },
-      numberNeeded: 2,
-      numberAllowedToParticipateAtSameTime: 1,
-      notificationDelay: 90,
-    }]),
-    description: 'While visiting a place of worship, create a half half photo.',
-    notificationText: 'View this and other available experiences',
-    callbacks: [{
-      trigger: '(cb.numberOfSubmissions() % 2) === 0',
-      function: halfhalfRespawnAndNotify('A Religious Architecture photo completed','View the photo').toString()
-    }]
-  },
-  halfhalf_sunset_knowsDTR: {
-    _id: Random.id(),
-    name: 'Sunset Together',
-    group: 'DTR',
-    participateTemplate: 'halfhalfParticipate',
-    resultsTemplate: 'halfhalfResults',
-    contributionTypes: addStaticAffordanceToNeeds('knowsDTR', [{
-      // needName MUST have structure "My Need Name XYZ"
-      needName: 'Sunset Together 1',
-      notificationSubject: 'Can you see the sunset?',
-      notificationText: 'Share an experience with others who are also watching the sunset',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.sunset),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'What does the <span style="color: #0351ff">sunset</span> look like where you are? Find a good view; then, take a photo, with your hands outstretched towards the sun.',
-        exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-sunset-heart.jpg'
-      },
-      numberNeeded: 2,
-      numberAllowedToParticipateAtSameTime: 1,
-      notificationDelay: 1,
-    }]),
-    description: 'While looking up at the sunset, create a half half photo.',
-    notificationText: 'View this and other available experiences',
-    callbacks: [{
-      trigger: '(cb.numberOfSubmissions() % 2) === 0',
-      function: halfhalfRespawnAndNotify('A Sunset Together photo completed','View the photo').toString()
-    }]
-  },
-  halfhalf_asian_knowsDTR: {
-    _id: Random.id(),
-    name: 'Eating with Chopsticks',
-    group: 'DTR',
-    participateTemplate: 'halfhalfParticipate',
-    resultsTemplate: 'halfhalfResults',
-    contributionTypes: addStaticAffordanceToNeeds('knowsDTR', [{
-      // needName MUST have structure "My Need Name XYZ"
-      needName: 'Eating with Chopsticks 1',
-      notificationSubject: 'Eating at an asian restaurant?',
-      notificationText: 'Share an experience with others who are also eating asian food',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.eating_with_chopsticks),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'Are you eating <span style="color: #0351ff">asian food</span> right now? Take a photo of what you are eating, <span style="color: #0351ff">holding your chopsticks.</span>',
-        exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-holding-chopsticks.jpg'
-      },
-      numberNeeded: 2,
-      numberAllowedToParticipateAtSameTime: 1,
-      notificationDelay: 90
-    }]),
-    description: 'While eating asian food, create a half half photo.',
-    notificationText: 'View this and other available experiences',
-    callbacks: [{
-      trigger: '(cb.numberOfSubmissions() % 2) === 0',
-      function: halfhalfRespawnAndNotify('An Eating with Chopsticks photo completed','View the photo').toString()
-    }]
-  },
-  halfhalf_books_knowsDTR: {
-    _id: Random.id(),
-    name: 'Book Buddies',
-    group: 'DTR',
-    participateTemplate: 'halfhalfParticipate',
-    resultsTemplate: 'halfhalfResults',
-    contributionTypes: addStaticAffordanceToNeeds('knowsDTR', [{
-      // needName MUST have structure "My Need Name XYZ"
-      needName: 'Book Buddies 1',
-      notificationSubject: 'Are you at a library?',
-      notificationText: 'Share an experience with others who are also at the library',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.library),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'Sorry to interrupt your <span style="color: #0351ff">reading</span>! Find the nearest book, and take a photo <span style="color: #0351ff">holding up the book to your face.</span>',
-        exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-book-face.jpg'
-      },
-      numberNeeded: 2,
-      numberAllowedToParticipateAtSameTime: 1,
-      notificationDelay: 90,
-    }]),
-    description: 'While reading a book, create a half half photo.',
-    notificationText: 'View this and other available experiences',
-    callbacks: [{
-      trigger: '(cb.numberOfSubmissions() % 2) === 0',
-      function: halfhalfRespawnAndNotify('A Book Buddies photo completed','View the photo').toString()
-    }]
-  },
-  halfhalf_leakmask_knowsDTR: {
-    _id: Random.id(),
-    name: 'Leaf Mask',
-    group: 'DTR',
-    participateTemplate: 'halfhalfParticipate',
-    resultsTemplate: 'halfhalfResults',
-    contributionTypes: addStaticAffordanceToNeeds('knowsDTR', [{
-      // needName MUST have structure "My Need Name XYZ"
-      needName: 'Leaf Mask 1',
-      notificationSubject: 'Are you at a park?',
-      notificationText: 'Share an experience with others who are also at a park',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.forest),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'Find a <span style="color: #0351ff">leaf in the park</span>. Take a photo of the <span style="color: #0351ff">leaf covering your face, like it was a mask.</span>',
-        exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-leaf-face.jpg'
-      },
-      numberNeeded: 2,
-      numberAllowedToParticipateAtSameTime: 1,
-      notificationDelay: 90
-    }]),
-    description: 'While in the park, create a half half photo.',
-    notificationText: 'View this and other available experiences',
-    callbacks: [{
-      trigger: '(cb.numberOfSubmissions() % 2) === 0',
-      function: halfhalfRespawnAndNotify('A Feet to the trees photo completed','View the photo').toString()
-    }]
-  },
-  halfhalf_puddles_knowsDTR: {
-    _id: Random.id(),
-    name: 'Puddle Feet',
-    group: 'DTR',
-    participateTemplate: 'halfhalfParticipate',
-    resultsTemplate: 'halfhalfResults',
-    contributionTypes: addStaticAffordanceToNeeds('knowsDTR', [{
-      // needName MUST have structure "My Need Name XYZ"
-      needName: 'Puddle Feet 1',
-      notificationSubject: 'Are you outside while its raining?',
-      notificationText: 'Share an experience with others who are enjoying or enduring the rain',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.rainy),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'Is it <span style="color: #0351ff">raining</span> today? Find a <span style="color: #0351ff">puddle</span> on the ground. Take a photo of yourself, stomping on the puddle!',
-        exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-puddle-feet.jpg'
-      },
-      numberNeeded: 2,
-      numberAllowedToParticipateAtSameTime: 1,
-      notificationDelay: 1,
-    }]),
-    description: 'With the puddles on a rainy day, create a half half photo.',
-    notificationText: 'View this and other available experiences',
-    callbacks: [{
-      trigger: '(cb.numberOfSubmissions() % 2) === 0',
-      function: halfhalfRespawnAndNotify('A "Puddle Feet" photo completed','View the photo').toString()
-    }]
-  },
-  halfhalf_coffeesideoflaughs_knowsDTR: {
-    _id: Random.id(),
-    name: "Coffee with a side of Laughs",
-    group: 'DTR',
-    participateTemplate: 'halfhalfParticipate',
-    resultsTemplate: 'halfhalfResults',
-    contributionTypes: addStaticAffordanceToNeeds('knowsDTR', [{
-      // needName MUST have structure "My Need Name XYZ"
-      needName: "Coffee with a side of Laughs 1",
-      notificationSubject: 'Inside a coffee shop?',
-      notificationText: 'Share an experience with others who are also at a coffee shop',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.coffee), // any place that has cups (cafes + bars + restaurants)
-        number: '1'
-      },
-      toPass: {
-        instruction: `Do you have <span style="color: #0351ff">a cup or glass</span> you are drinking? Take a photo with it in the middle of the picture. You can even try to <span style="color: #0351ff">pour some extra "cream"</span> into it too!`,
-        exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-teasing-lotion-in-a-cup.jpg'
-      },
-      numberNeeded: 2,
-      notificationDelay: 90,
-    }]),
-    description: 'While drinking coffee at a cafe, create a half half photo.',
-    notificationText: 'View this and other available experiences',
-    callbacks: [{
-      trigger: '(cb.numberOfSubmissions() % 2) === 0',
-      function: halfhalfRespawnAndNotify("A 'Coffee with a side of Laughs' photo completed",'View the photo').toString()
-    }]
-  },
-  halfhalf_bigbites_knowsDTR: {
-    _id: Random.id(),
-    name: "Big Bites",
-    group: 'DTR',
-    participateTemplate: 'halfhalfParticipate',
-    resultsTemplate: 'halfhalfResults',
-    contributionTypes: addStaticAffordanceToNeeds('knowsDTR', [{
-      needName: "Big Bites 1", // Any restaurant that would serve something you'd eat with your hands (burrito, tacos, hotdogs, sandwiches, wraps, burgers, tradamerican, newamerican )
-      notificationSubject: 'Eating at a restaurant?',
-      notificationText: 'Share an experience with others who are enjoying big bites of their meal',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.big_bite_restaurant),
-        number: '1'
-      },
-      toPass: {
-        instruction: `Are you <span style="color: #0351ff">eating food that would require a big bite</span> right now? Take a photo of yourself <span style="color: #0351ff">holding up your food</span> to the middle of the screen.`,
-        exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-big-bite.jpg'
-      },
-      numberNeeded: 2,
-      numberAllowedToParticipateAtSameTime: 1,
-      notificationDelay: 90, // https://www.quora.com/Whats-the-average-time-that-customers-wait-between-entering-a-restaurant-and-getting-served
-    }]),
-    description: 'While eating some non-trivially sized food, create a half half photo.',
-    notificationText: 'View this and other available experiences',
-    callbacks: [{
-      trigger: '(cb.numberOfSubmissions() % 2) === 0',
-      function: halfhalfRespawnAndNotify("A 'Big Bites' photo completed",'View the photo').toString()
-    }]
-  },
-};
-
-let CHI20_Olin_EXPERIENCES =  {
-  storytime_knowsOlin: createStorytimeImproved('Olin', 1), // version 1 starts with bars
-  halfhalf_sunny_knowsOlin: {
-    _id: Random.id(),
-    name: 'Hand Silhouette',
-    group: 'Olin',
-    participateTemplate: 'halfhalfParticipate',
-    resultsTemplate: 'halfhalfResults',
-    contributionTypes: addStaticAffordanceToNeeds('knowsOlin', [{
-      // needName MUST have structure "My Need Name XYZ"
-      needName: 'Hand Silhouette 1',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.sunny),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'Is the <span style="color: #0351ff">weather clear and sunny</span> where you are? Take a photo, <span style="color: #0351ff">holding your hand towards the sky, covering the sun.</span>',
-        exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-hands-in-front.jpg'
-      },
-      numberNeeded: 2,
-      numberAllowedToParticipateAtSameTime: 1,
-      notificationDelay: 1,
-    }]),
-    description: 'Use the sun to make a silhouette of your hand',
-    notificationText: 'View this and other available experiences',
-    callbacks: [{
-      trigger: '(cb.numberOfSubmissions() % 2) === 0',
-      function: halfhalfRespawnAndNotify('A hand silhouette was completed','View the photo').toString()
-    }]
-  },
-  halfhalf_grocery_knowsOlin: {
-    _id: Random.id(),
-    name: 'Grocery Buddies',
-    group: 'Olin',
-    participateTemplate: 'halfhalfParticipate',
-    resultsTemplate: 'halfhalfResults',
-    contributionTypes: addStaticAffordanceToNeeds('knowsOlin', [{
-      // needName MUST have structure "My Need Name XYZ"
-      needName: 'Grocery Buddies 1',
-      notificationSubject: 'Inside a grocery store?',
-      notificationText: 'Share an experience with others who are also grocery shopping',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.grocery),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'Are you at the <span style="color: #0351ff">grocery store</span>? Take a photo, <span style="color: #0351ff">holding a fruit or vegetable</span> outstretched with your hands.',
-        exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-fruit-in-hand.jpg'
-      },
-      numberNeeded: 2,
-      numberAllowedToParticipateAtSameTime: 1,
-      notificationDelay: 90,
-    }]),
-    description: 'While shopping for groceries, create a half half photo.',
-    notificationText: 'View this and other available experiences',
-    callbacks: [{
-      trigger: '(cb.numberOfSubmissions() % 2) === 0',
-      function: halfhalfRespawnAndNotify('A Grocery Buddies photo completed','View the photo').toString()
-    }]
-  },
-  halfhalf_bar_knowsOlin: {
-    _id: Random.id(),
-    name: 'Cheers',
-    group: 'Olin',
-    participateTemplate: 'halfhalfParticipate',
-    resultsTemplate: 'halfhalfResults',
-    contributionTypes: addStaticAffordanceToNeeds('knowsOlin', [{
-      // needName MUST have structure "My Need Name XYZ"
-      needName: 'Cheers 1',
-      notificationSubject: 'Drinking at a bar?',
-      notificationText: 'Share an experience with others who are also drinking at a bar',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.bar),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'What are you <span style="color: #0351ff">drinking at the bar</span>? Take a photo, while <span style="color: #0351ff">raising your glass or bottle</span> in front of you.',
-        exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-cheers.jpg'
-      },
-      numberNeeded: 2,
-      numberAllowedToParticipateAtSameTime: 1,
-      notificationDelay: 90
-    }]),
-    description: 'While enjoying your drink, create a half half photo.',
-    notificationText: 'View this and other available experiences',
-    callbacks: [{
-      trigger: '(cb.numberOfSubmissions() % 2) === 0',
-      function: halfhalfRespawnAndNotify('A Cheers photo completed','View the photo').toString()
-    }]
-  },
-  halfhalf_religious_knowsOlin: {
-    _id: Random.id(),
-    name: 'Religious Architecture',
-    group: 'Olin',
-    participateTemplate: 'halfhalfParticipate',
-    resultsTemplate: 'halfhalfResults',
-    contributionTypes: addStaticAffordanceToNeeds('knowsOlin', [{
-      // needName MUST have structure "My Need Name XYZ"
-      needName: 'Religious Architecture 1',
-      notificationSubject: 'Visiting a place of worship?',
-      notificationText: 'Share an experience with others who are also visiting a place of worship',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.castle),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'Do you notice the <span style="color: #0351ff">details of the religious building</span> near you? Do so now, by outstretching your hand and pointing out of the elements that stick out to you most.',
-        exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-religious-building.jpg'
-      },
-      numberNeeded: 2,
-      numberAllowedToParticipateAtSameTime: 1,
-      notificationDelay: 90,
-    }]),
-    description: 'While visiting a place of worship, create a half half photo.',
-    notificationText: 'View this and other available experiences',
-    callbacks: [{
-      trigger: '(cb.numberOfSubmissions() % 2) === 0',
-      function: halfhalfRespawnAndNotify('A Religious Architecture photo completed','View the photo').toString()
-    }]
-  },
-  halfhalf_sunset_knowsOlin: {
-    _id: Random.id(),
-    name: 'Sunset Together',
-    group: 'Olin',
-    participateTemplate: 'halfhalfParticipate',
-    resultsTemplate: 'halfhalfResults',
-    contributionTypes: addStaticAffordanceToNeeds('knowsOlin', [{
-      // needName MUST have structure "My Need Name XYZ"
-      needName: 'Sunset Together 1',
-      notificationSubject: 'Can you see the sunset?',
-      notificationText: 'Share an experience with others who are also watching the sunset',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.sunset),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'What does the <span style="color: #0351ff">sunset</span> look like where you are? Find a good view; then, take a photo, with your hands outstretched towards the sun.',
-        exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-sunset-heart.jpg'
-      },
-      numberNeeded: 2,
-      numberAllowedToParticipateAtSameTime: 1,
-      notificationDelay: 1,
-    }]),
-    description: 'While looking up at the sunset, create a half half photo.',
-    notificationText: 'View this and other available experiences',
-    callbacks: [{
-      trigger: '(cb.numberOfSubmissions() % 2) === 0',
-      function: halfhalfRespawnAndNotify('A Sunset Together photo completed','View the photo').toString()
-    }]
-  },
-  halfhalf_asian_knowsOlin: {
-    _id: Random.id(),
-    name: 'Eating with Chopsticks',
-    group: 'Olin',
-    participateTemplate: 'halfhalfParticipate',
-    resultsTemplate: 'halfhalfResults',
-    contributionTypes: addStaticAffordanceToNeeds('knowsOlin', [{
-      // needName MUST have structure "My Need Name XYZ"
-      needName: 'Eating with Chopsticks 1',
-      notificationSubject: 'Eating at an asian restaurant?',
-      notificationText: 'Share an experience with others who are also eating asian food',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.eating_with_chopsticks),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'Are you eating <span style="color: #0351ff">asian food</span> right now? Take a photo of what you are eating, <span style="color: #0351ff">holding your chopsticks.</span>',
-        exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-holding-chopsticks.jpg'
-      },
-      numberNeeded: 2,
-      numberAllowedToParticipateAtSameTime: 1,
-      notificationDelay: 90
-    }]),
-    description: 'While eating asian food, create a half half photo.',
-    notificationText: 'View this and other available experiences',
-    callbacks: [{
-      trigger: '(cb.numberOfSubmissions() % 2) === 0',
-      function: halfhalfRespawnAndNotify('An Eating with Chopsticks photo completed','View the photo').toString()
-    }]
-  },
-  halfhalf_books_knowsOlin: {
-    _id: Random.id(),
-    name: 'Book Buddies',
-    group: 'Olin',
-    participateTemplate: 'halfhalfParticipate',
-    resultsTemplate: 'halfhalfResults',
-    contributionTypes: addStaticAffordanceToNeeds('knowsOlin', [{
-      // needName MUST have structure "My Need Name XYZ"
-      needName: 'Book Buddies 1',
-      notificationSubject: 'Are you at a library?',
-      notificationText: 'Share an experience with others who are also at the library',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.library),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'Sorry to interrupt your <span style="color: #0351ff">reading</span>! Find the nearest book, and take a photo <span style="color: #0351ff">holding up the book to your face.</span>',
-        exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-book-face.jpg'
-      },
-      numberNeeded: 2,
-      numberAllowedToParticipateAtSameTime: 1,
-      notificationDelay: 90,
-    }]),
-    description: 'While reading a book, create a half half photo.',
-    notificationText: 'View this and other available experiences',
-    callbacks: [{
-      trigger: '(cb.numberOfSubmissions() % 2) === 0',
-      function: halfhalfRespawnAndNotify('A Book Buddies photo completed','View the photo').toString()
-    }]
-  },
-  halfhalf_leakmask_knowsOlin: {
-    _id: Random.id(),
-    name: 'Leaf Mask',
-    group: 'Olin',
-    participateTemplate: 'halfhalfParticipate',
-    resultsTemplate: 'halfhalfResults',
-    contributionTypes: addStaticAffordanceToNeeds('knowsOlin', [{
-      // needName MUST have structure "My Need Name XYZ"
-      needName: 'Leaf Mask 1',
-      notificationSubject: 'Are you at a park?',
-      notificationText: 'Share an experience with others who are also at a park',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.forest),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'Find a <span style="color: #0351ff">leaf in the park</span>. Take a photo of the <span style="color: #0351ff">leaf covering your face, like it was a mask.</span>',
-        exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-leaf-face.jpg'
-      },
-      numberNeeded: 2,
-      numberAllowedToParticipateAtSameTime: 1,
-      notificationDelay: 90
-    }]),
-    description: 'While in the park, create a half half photo.',
-    notificationText: 'View this and other available experiences',
-    callbacks: [{
-      trigger: '(cb.numberOfSubmissions() % 2) === 0',
-      function: halfhalfRespawnAndNotify('A Feet to the trees photo completed','View the photo').toString()
-    }]
-  },
-  halfhalf_puddles_knowsOlin: {
-    _id: Random.id(),
-    name: 'Puddle Feet',
-    group: 'Olin',
-    participateTemplate: 'halfhalfParticipate',
-    resultsTemplate: 'halfhalfResults',
-    contributionTypes: addStaticAffordanceToNeeds('knowsOlin', [{
-      // needName MUST have structure "My Need Name XYZ"
-      needName: 'Puddle Feet 1',
-      notificationSubject: 'Are you outside while its raining?',
-      notificationText: 'Share an experience with others who are enjoying or enduring the rain',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.rainy),
-        number: '1'
-      },
-      toPass: {
-        instruction: 'Is it <span style="color: #0351ff">raining</span> today? Find a <span style="color: #0351ff">puddle</span> on the ground. Take a photo of yourself, stomping on the puddle!',
-        exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-puddle-feet.jpg'
-      },
-      numberNeeded: 2,
-      numberAllowedToParticipateAtSameTime: 1,
-      notificationDelay: 1,
-    }]),
-    description: 'With the puddles on a rainy day, create a half half photo.',
-    notificationText: 'View this and other available experiences',
-    callbacks: [{
-      trigger: '(cb.numberOfSubmissions() % 2) === 0',
-      function: halfhalfRespawnAndNotify('A "Puddle Feet" photo completed','View the photo').toString()
-    }]
-  },
-  halfhalf_coffeesideoflaughs_knowsOlin: {
-    _id: Random.id(),
-    name: "Coffee with a side of Laughs",
-    group: 'Olin',
-    participateTemplate: 'halfhalfParticipate',
-    resultsTemplate: 'halfhalfResults',
-    contributionTypes: addStaticAffordanceToNeeds('knowsOlin', [{
-      // needName MUST have structure "My Need Name XYZ"
-      needName: "Coffee with a side of Laughs 1",
-      notificationSubject: 'Inside a coffee shop?',
-      notificationText: 'Share an experience with others who are also at a coffee shop',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.coffee), // any place that has cups (cafes + bars + restaurants)
-        number: '1'
-      },
-      toPass: {
-        instruction: `Do you have <span style="color: #0351ff">a cup or glass</span> you are drinking? Take a photo with it in the middle of the picture. You can even try to <span style="color: #0351ff">pour some extra "cream"</span> into it too!`,
-        exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-teasing-lotion-in-a-cup.jpg'
-      },
-      numberNeeded: 2,
-      notificationDelay: 90,
-    }]),
-    description: 'While drinking coffee at a cafe, create a half half photo.',
-    notificationText: 'View this and other available experiences',
-    callbacks: [{
-      trigger: '(cb.numberOfSubmissions() % 2) === 0',
-      function: halfhalfRespawnAndNotify("A 'Coffee with a side of Laughs' photo completed",'View the photo').toString()
-    }]
-  },
-  halfhalf_bigbites_knowsOlin: {
-    _id: Random.id(),
-    name: "Big Bites",
-    group: 'Olin',
-    participateTemplate: 'halfhalfParticipate',
-    resultsTemplate: 'halfhalfResults',
-    contributionTypes: addStaticAffordanceToNeeds('knowsOlin', [{
-      needName: "Big Bites 1", // Any restaurant that would serve something you'd eat with your hands (burrito, tacos, hotdogs, sandwiches, wraps, burgers, tradamerican, newamerican )
-      notificationSubject: 'Eating at a restaurant?',
-      notificationText: 'Share an experience with others who are enjoying big bites of their meal',
-      situation: {
-        detector: getDetectorUniqueKey(DETECTORS.big_bite_restaurant),
-        number: '1'
-      },
-      toPass: {
-        instruction: `Are you <span style="color: #0351ff">eating food that would require a big bite</span> right now? Take a photo of yourself <span style="color: #0351ff">holding up your food</span> to the middle of the screen.`,
-        exampleImage: 'https://s3.us-east-2.amazonaws.com/ce-platform/oce-example-images/half-half-embodied-mimicry-big-bite.jpg'
-      },
-      numberNeeded: 2,
-      numberAllowedToParticipateAtSameTime: 1,
-      notificationDelay: 90, // https://www.quora.com/Whats-the-average-time-that-customers-wait-between-entering-a-restaurant-and-getting-served
-    }]),
-    description: 'While eating some non-trivially sized food, create a half half photo.',
-    notificationText: 'View this and other available experiences',
-    callbacks: [{
-      trigger: '(cb.numberOfSubmissions() % 2) === 0',
-      function: halfhalfRespawnAndNotify("A 'Big Bites' photo completed",'View the photo').toString()
-    }]
-  },
-};
+import CHI20_DTR_EXPERIENCES from "./summer19study/dtr_experiences";
+import CHI20_Olin_EXPERIENCES from "./summer19study/olin_experiences";
 
 export const CONSTANTS = {
   'LOCATIONS': LOCATIONS,
@@ -3032,44 +2165,3 @@ export const CONSTANTS = {
   // 'EXPERIENCES': CHI20_DTR_EXPERIENCES,
   'DETECTORS': DETECTORS
 };
-
-// Meteor.call('locations.updateUserLocationAndAffordances', {
-//   uid: Accounts.findUserByUsername('b@gmail.com')._id,
-//   lat: 42.054902,  //lakefill
-//   lng: -87.670197
-// });
-// Meteor.call('locations.updateUserLocationAndAffordances', {
-//   uid: Accounts.findUserByUsername('c@gmail.com')._id,
-//   lat: 42.056975, //ford
-//   lng:  -87.676575
-// });
-// Meteor.call('locations.updateUserLocationAndAffordances', {
-//   uid: Accounts.findUserByUsername('d@gmail.com')._id,
-//   lat: 42.059273, //garage
-//   lng: -87.673794
-// });
-// Meteor.call('locations.updateUserLocationAndAffordances', {
-//   uid: Accounts.findUserByUsername('e@gmail.com')._id,
-//   lat: 42.044314,  //nevins
-//   lng: -87.682157
-// });
-//
-// Meteor.call('locations.updateUserLocationAndAffordances', {
-//   uid: Accounts.findUserByUsername('g@gmail.com')._id,
-//   lat: 42.044314,  //nevins
-//   lng: -87.682157
-// });
-// Meteor.call('locations.updateUserLocationAndAffordances', {
-//   uid: Accounts.findUserByUsername('h@gmail.com')._id,
-//   lat: 42.045398,  //pubs
-//   lng: -87.682431
-// });
-// Meteor.call('locations.updateUserLocationAndAffordances', {
-//   uid: Accounts.findUserByUsername('i@gmail.com')._id,
-//   lat: 42.047621, //grocery, whole foods
-//   lng: -87.679488
-// });
-// Meteor.call('locations.updateUserLocationAndAffordances', {
-//   uid: Accounts.findUserByUsername('j@gmail.com')._id,
-//   lat: 42.042617, //beach
-//   lng: -87.671474
